moved mock to its own file, clean up
