+++ TKeazirian.HTTPServer.Tests/Handler/RouterTests.cs
@@ -10,23 +10,6 @@ using TKeazirian.HTTPServer.Response;
 
 public class RouterTests
 {
-    private class MockHandler : Handler
-    {
-        public override List<string> AllowedHttpMethods()
-        {
-            return new List<string>() { "GET" };
-        }
-
-        public override Response HandleResponse(Request request)
-        {
-            return new Response(
-                "HTTP/1.1 200 OK\r\n",
-                "Content-Type: text/plain\r\nContent-Length: 9\r\n\r\n",
-                "Mock body"
-            );
-        }
-    }
-
     [Fact]
     public void RouterRoutesToExpectedHandler()
     {
